Adding the following folders:
trunk: main folder to hold the OpenDatcom source for the main development of the project.

branches: folder to contain separate branches for developers to make additions and add separate functionality that could possibly break the build at first.  Once a functionality is implemented and integrated the branch will be merged back in with the trunk.    

tags: folder to contain separate releases of OpenDatcom (eg. OpenDatcom v1.0).